Import Boost 1.58.0

+++ libs/geometry/test/algorithms/overlay/self_intersection_points.cpp
@@ -47,7 +47,7 @@ static void test_self_intersection_points(std::string const& case_id,
             int expected_count,
             Geometry const& geometry,
             bool check_has_intersections,
-            double precision = 0.001)
+            double /*precision*/ = 0.001)
 {
     typedef typename bg::point_type<Geometry>::type point_type;
     //typedef typename bg::rescale_policy_type<point_type>::type rescale_policy_type;
